Refactor: Move AI data models to dedicated package
This commit relocates several data classes used for AI prompt context and generation results into a new `core.ai.models` package. This improves code organization and separation of concerns.

**Moved classes:**
- `ActConclusionContext`
- `ChapterConclusionContext`
- `SagaEndCreditsContext`
- `SagaGenerationContext`

**Other changes:**
- Deleted `CharacterFraming.kt` as it was unused.
- Updated import statements across various files to reflect the new package structure for the moved models.
- Minor cleanup of unused imports in `EmotionalPrompt.kt`, `SagaHistoryUseCase.kt`, and `SagaHistoryUseCaseImpl.kt`.
- Corrected package for `GetInputSuggestionsUseCaseImpl`.
- Updated `SuggestionPrompts.kt` to use the moved `SuggestionGen` model.
