+++ app/src/main/java/com/ilustris/sagai/core/ai/models/SagaGenerationContext.kt
@@ -1,0 +1,8 @@
+package com.ilustris.sagai.core.ai.models
+
+import com.ilustris.sagai.features.newsaga.data.model.SagaForm
+
+data class SagaGenerationContext(
+    val sagaSetup: SagaForm,
+    val initialPlayerInteractionLog: String,
+)
