+++ app/src/main/java/com/ilustris/sagai/core/ai/prompts/SuggestionPrompts.kt
@@ -1,15 +1,13 @@
 package com.ilustris.sagai.core.ai.prompts
 
 import com.ilustris.sagai.core.utils.formatToString
-import com.ilustris.sagai.core.utils.toJsonFormat
 import com.ilustris.sagai.core.utils.toJsonFormatExcludingFields
 import com.ilustris.sagai.core.utils.toJsonMap
 import com.ilustris.sagai.features.characters.data.model.Character
 import com.ilustris.sagai.features.home.data.model.Saga
+import com.ilustris.sagai.features.saga.chat.data.model.SuggestionGen
 import com.ilustris.sagai.features.saga.chat.domain.model.MessageContent
-import com.ilustris.sagai.features.saga.chat.domain.model.SuggestionsReponse
 import com.ilustris.sagai.features.saga.chat.domain.model.joinMessage
-import com.ilustris.sagai.features.saga.chat.domain.usecase.model.SuggestionGen
 
 object SuggestionPrompts {
     fun generateSuggestionsPrompt(
@@ -20,12 +18,12 @@ object SuggestionPrompts {
         val historyFormatted = chatHistory.joinToString { it.joinMessage().formatToString() }
 
         return """
-                                                You are an AI assistant for a text-based role-playing game.
-                                                Your goal is to provide three concise, creative, contextually relevant, and actionable input suggestions for the player,
-                                                who is currently embodying '${character.name}'.
+                                                            You are an AI assistant for a text-based role-playing game.
+                                                            Your goal is to provide three concise, creative, contextually relevant, and actionable input suggestions for the player,
+                                                            who is currently embodying '${character.name}'.
 
-                                                Saga Context:
-                                                ${saga.toJsonFormatExcludingFields(
+                                                            Saga Context:
+                                                            ${saga.toJsonFormatExcludingFields(
             listOf(
                 "currentActId",
                 "isEnded",
@@ -41,8 +39,8 @@ object SuggestionPrompts {
             ),
         )}
 
-                                                Character context:
-                                                ${character.toJsonFormatExcludingFields(
+                                                            Character context:
+                                                            ${character.toJsonFormatExcludingFields(
             listOf(
                 "hexColor",
                 "image",
@@ -59,39 +57,39 @@ object SuggestionPrompts {
             ),
         )}
 
-                                                Recent Chat History (for immediate context, newest messages are most relevant):
-                                                $historyFormatted
+                                                            Recent Chat History (for immediate context, newest messages are most relevant):
+                                                            $historyFormatted
 
-                                                Task:
-                                                Based on all the information above, generate exactly 3 distinct input suggestions for the player controlling '${character.name}'.
-                                                Each suggestion must include the `text` of the suggestion and its `type`.
-                                                The suggestions should guide the player's next interaction.
+                                                            Task:
+                                                            Based on all the information above, generate exactly 3 distinct input suggestions for the player controlling '${character.name}'.
+                                                            Each suggestion must include the `text` of the suggestion and its `type`.
+                                                            The suggestions should guide the player's next interaction.
 
-                                                Allowed `type` values (from your game's SenderType system):
-                                                - "USER": For dialogue spoken by '${character.name}'. 
-                                                   The text should be in first person a direct speech.
-                                                - "THOUGHT": For an internal thought of '${character.name}'. 
-                                                    The text should be a internal thought, not a direct speech.
-                                                - "ACTION": For a physical action '${character.name}' performs. 
-                                                The text should be a concise description of the action in the present tense (e.g., "Open the ancient chest", "Look through the keyhole", "Follow the mysterious footprints"). Avoid phrases like "I will..." or "Character will...".
-                                                - "NARRATOR": To suggest a focus on an environmental detail or a subtle observation that '${character.name}' might notice or want to investigate further. The text should be a statement of that observation (e.g., "The dusty tome on the lectern seems to call out.", "A faint scratching sound comes from behind the wall."). This type prompts the player to explore or inquire.
+                                                            Allowed `type` values (from your game's SenderType system):
+                                                            - "USER": For dialogue spoken by '${character.name}'. 
+                                                               The text should be in first person a direct speech.
+                                                            - "THOUGHT": For an internal thought of '${character.name}'. 
+                                                                The text should be a internal thought, not a direct speech.
+                                                            - "ACTION": For a physical action '${character.name}' performs. 
+                                                            The text should be a concise description of the action in the present tense (e.g., "Open the ancient chest", "Look through the keyhole", "Follow the mysterious footprints"). Avoid phrases like "I will..." or "Character will...".
+                                                            - "NARRATOR": To suggest a focus on an environmental detail or a subtle observation that '${character.name}' might notice or want to investigate further. The text should be a statement of that observation (e.g., "The dusty tome on the lectern seems to call out.", "A faint scratching sound comes from behind the wall."). This type prompts the player to explore or inquire.
 
-                                                Ensure suggestions are creative, contextually relevant, and brief, suitable for quick selection.
+                                                            Ensure suggestions are creative, contextually relevant, and brief, suitable for quick selection.
 
-                                                Output Format:
-                                                Return ONLY a valid JSON array of objects. Each object must have two fields:
-                                                1.  `text`: (string) The suggestion text.
-                                                2.  `type`: (string) One of the allowed SenderType values: "USER", "THOUGHT", "ACTION", "NARRATOR".
+                                                            Output Format:
+                                                            Return ONLY a valid JSON array of objects. Each object must have two fields:
+                                                            1.  `text`: (string) The suggestion text.
+                                                            2.  `type`: (string) One of the allowed SenderType values: "USER", "THOUGHT", "ACTION", "NARRATOR".
 
-                                                Do not include any other text, explanations, or markdown.
+                                                            Do not include any other text, explanations, or markdown.
 
-                                                Example Output:
-                                                [
-                                                  {"text": "What secrets does this map hold?", "type": "THOUGHT"},
-                                                  {"text": "Unroll the weathered map.", "type": "ACTION"},
-                                                  {"text": "The symbols on the map pulse with a faint, ethereal glow.", "type": "NARRATOR"}
-                                                ]
-                                                ${OutputRules.outputRule(toJsonMap(SuggestionGen::class.java))}
+                                                            Example Output:
+                                                            [
+                                                              {"text": "What secrets does this map hold?", "type": "THOUGHT"},
+                                                              {"text": "Unroll the weathered map.", "type": "ACTION"},
+                                                              {"text": "The symbols on the map pulse with a faint, ethereal glow.", "type": "NARRATOR"}
+                                                            ]
+                                                            ${OutputRules.outputRule(toJsonMap(SuggestionGen::class.java))}
             """.trimIndent()
     }
 }
