+++ app/src/main/java/com/ilustris/sagai/features/saga/chat/data/model/SuggestionGen.kt
@@ -1,2 +1,7 @@
-package com.ilustris.sagai.features.saga.chat.data.model 
+package com.ilustris.sagai.features.saga.chat.data.model
 
+import com.ilustris.sagai.features.saga.chat.domain.model.Suggestion
+
+data class SuggestionGen(
+    val suggestions: List<Suggestion>,
+)
