+++ app/src/main/java/com/ilustris/sagai/features/saga/chat/data/usecase/GetInputSuggestionsUseCaseImpl.kt
@@ -1,4 +1,4 @@
-package com.ilustris.sagai.features.saga.chat.domain.usecase
+package com.ilustris.sagai.features.saga.chat.data.usecase
 
 import android.util.Log
 import com.ilustris.sagai.core.ai.GemmaClient
@@ -8,9 +8,10 @@ import com.ilustris.sagai.core.data.asError
 import com.ilustris.sagai.core.data.asSuccess
 import com.ilustris.sagai.features.characters.data.model.Character
 import com.ilustris.sagai.features.home.data.model.Saga
+import com.ilustris.sagai.features.saga.chat.data.model.SuggestionGen
 import com.ilustris.sagai.features.saga.chat.domain.model.MessageContent
 import com.ilustris.sagai.features.saga.chat.domain.model.Suggestion
-import com.ilustris.sagai.features.saga.chat.domain.usecase.model.SuggestionGen
+import com.ilustris.sagai.features.saga.chat.domain.usecase.GetInputSuggestionsUseCase
 import javax.inject.Inject
 
 class GetInputSuggestionsUseCaseImpl
